[Silver III] Title: 주유소, Time: 360 ms, Memory: 30928 KB -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -6,20 +6,20 @@ public class Main {
     	StringTokenizer st;
     	int n = Integer.parseInt(br.readLine());
     	
-    	long[] dis = new long[n-1];
-    	long[] cost = new long[n];
+    	int[] dis = new int[n-1];
+    	int[] cost = new int[n];
     	st = new StringTokenizer(br.readLine());
     	for(int i = 0; i < n-1; i++) {
-    		dis[i] = Long.parseLong(st.nextToken());
+    		dis[i] = Integer.parseInt(st.nextToken());
     	}
     	
     	st = new StringTokenizer(br.readLine());
     	for(int i = 0; i < n; i++) {
-    		cost[i] = Long.parseLong(st.nextToken());
+    		cost[i] = Integer.parseInt(st.nextToken());
     	}
     	
-    	long sum = 0;
-    	long min = cost[0];
+    	int sum = 0;
+    	int min = cost[0];
     	for(int i = 0; i < n-1; i++) {
     		if(cost[i] < min) min = cost[i];
     		sum += min * dis[i];
